speed related optimizations

+++ dmx.h
@@ -2,7 +2,7 @@
 #define __DMX_H
 
 #define MAX_DMX_CHANNELS 32
-volatile uint8_t dmxChannels[MAX_DMX_CHANNELS];
+volatile uint32_t dmxChannels[MAX_DMX_CHANNELS];
 
 
 #endif
